Started work on Stress feature
Added a logging feature that outputs different log levels depending on global currentLogLevel variable.

Added commands to register as a StressCharacter, from which point on stress is tracked for this character.
Added a command to increase stress, which currently simply updates the persisted stress number on the character.

Wrote the start of a State manager for the Stress State